chore(mcp): move openai usage to new architecture
